+++ src-tauri/src/models/recipe_detail.rs
@@ -4,14 +4,14 @@ use sqlx::prelude::FromRow;
 #[derive(Debug, Deserialize, FromRow, Serialize, Clone)]
 pub struct RecipeDetail {
     pub id: String,
-    pub input_amount: f64,
+    pub input_amount: i64,
     pub item_id: String,
     pub recipe_id: String,
 }
 
 #[derive(Debug, Deserialize, FromRow, Serialize, Clone)]
 pub struct RecipeDetailDto {
-    pub input_amount: f64,
+    pub input_amount: i64,
     pub item_id: String,
     pub recipe_id: Option<String>,
 }
@@ -19,7 +19,7 @@ pub struct RecipeDetailDto {
 #[derive(Debug, Deserialize, FromRow, Serialize, Clone)]
 pub struct RecipeDetailForRecipe {
     pub id: String,
-    pub input_amount: f64,
+    pub input_amount: i64,
     pub item_id: String,
     pub recipe_id: String,
     pub item_picture: String,
